Added blank design document
Hi Guys,

Ok, this is a template design document I FINALLY figured out how to get
onto GIT :).

There is a lot of information on here that we will not need, however,
this is an example of what a professional design document will have in
it. Generally before this document is completed, teams will submit a
project plan, which outlines the basic information about the game
(style, storyline, programming language, or anything that defines the
game.)The more spacific things get at these begining stages, the better
the pipeline will be - especially working in a team environment. So I
would suggest taking a read through this document quickly to see if you
have any ideas of what the game would look like in each of the headings
listed, and then start dreaming how the game would work. I will upload
a general project plan document as well (not nearly as crazy as this
one), which will be much more fun to fill in, and comment on each
others ideas. Thanks! Jaimie